FEAT: Implemented Repo and Other Changes to ChecklistStartVM

+++ app/src/main/java/com/example/grocerychecklist/ui/screen/checklist/ChecklistViewScreen.kt
@@ -127,7 +127,7 @@ fun ChecklistViewScreen(
                         variant = ChecklistItemComponentVariant.ChecklistItem,
                         category = item.category,
                         price = item.price,
-                        quantity = item.quantity,
+                        quantity = item.quantity.toDouble(),
                         measurement = item.measurement,
                     )
                 }
